Update Tabs Test Cases
