+++ src/test/java/seedu/address/ui/tab/FinanceTabTest.java
@@ -1,5 +1,4 @@
 package seedu.address.ui.tab;
 
 public class FinanceTabTest {
-
 }
